+++ R/00-NAMESPACE.R
@@ -1,10 +1,11 @@
 #' @import shiny
 #' @importFrom colourpicker colourInput updateColourInput
+#' @importFrom DataTools checkErrorNoNumericColumns checkWarningEmptyValues importDataUI
+#'  importDataServer remoteModelsUI remoteModelsServer tryCatchWithWarningsAndErrors
 #' @importFrom dplyr bind_cols bind_rows filter if_else mutate
 #' @importFrom DT DTOutput renderDT
 #' @importFrom graphics arrows axis box lines mtext par plot points rect text
 #' @importFrom grDevices dev.off pdf png recordPlot replayPlot rgb svg tiff
-#' @importFrom DataTools checkErrorNoNumericColumns checkWarningEmptyValues importDataUI importDataServer
 #' @importFrom jsonlite toJSON
 #' @importFrom magrittr %>%
 #' @importFrom mgcv Predict.matrix smoothCon s