+++ R/01-fitModel.R
@@ -6,44 +6,16 @@ utils::globalVariables(c("standardizedX", "Date3"))
 #' Get plotValues
 #'
 #' @param plotValues (list) list with all values required for the plot
-#' @param activeFile (character) name of the selected file
-#' @param activeFileData (data.frame) content of the selected file
-#' @param dataSelection (list) list of reactive selected columns of the
-#'  current file
 #' @param modelParameters (list) list of settings for the model
-getPlotValues <- function(plotValues, activeFile, activeFileData, dataSelection,
-                          modelParameters){
+fitModelWrapper <- function(plotValues,
+                            modelParameters){
   plotValues$modelData <- NULL
   plotValues$predictedData <- NULL
-
-  plotValues$activeFile <- activeFile
-  plotValues$activeFileData <- activeFileData
-  plotValues$dataSettings <- list(xColumns = dataSelection$xColumns(),
-                                  yColumns = dataSelection$yColumns(),
-                                  dataOutlier = dataSelection$dataOutlier())
 
   xSelection <- getSelection(plotValues$dataSettings$xColumns)
   ySelection <- getSelection(plotValues$dataSettings$yColumns)
 
-  plotValues$selectedData <-
-    activeFileData[, unlist(c(xSelection$colNames, ySelection$colNames))]
-
-  if (any(!sapply(plotValues$selectedData, is.numeric))) {
-    # transform to numeric
-    plotValues$selectedData <- toNumericCols(plotValues$selectedData)
-  }
-
-  if (!is.null(plotValues$selectedData) && any(is.na(plotValues$selectedData))) {
-    plotValues$selectedData <- na.omit(plotValues$selectedData)
-  }
-
   if (is.null(plotValues$selectedData)) return(plotValues)
-
-  plotValues$selectedData <- addColumnDataOutlier(
-    selectedData = plotValues$selectedData,
-    ySelection = ySelection,
-    dataOutlier = plotValues$dataSettings$dataOutlier
-  )
 
   prepData <- getPrepData(data = plotValues$selectedData,
                           xSelection = xSelection,
@@ -62,8 +34,10 @@ getPlotValues <- function(plotValues, activeFile, activeFileData, dataSelection,
 
   plotValues$predictedData <- predictData(modelData = plotValues$modelData$modelOutput,
                                prepData = prepData,
-                               smoothConst = plotValues$modelParameters$smoothConst)
+                               smoothConst = plotValues$modelParameters$smoothConst) %>%
+    tryCatchWithWarningsAndErrors(errorTitle = "Prediction failed", alertStyle = "shinyalert")
 
+  req(!is.null(plotValues$predictedData))
   plotValues$defaultXRange <- getRange(
     data = plotValues$selectedData[, unlist(xSelection$colNames),
                                    drop = FALSE],
@@ -99,8 +73,10 @@ getModelFit <- function(data,
 
     predictedData <- predictData(modelData = modelOutput,
                                  prepData = prepData,
-                                 smoothConst = modelParameters$smoothConst)
+                                 smoothConst = modelParameters$smoothConst) %>%
+      tryCatchWithWarningsAndErrors(errorTitle = "Prediction failed", alertStyle = "shinyalert")
 
+    req(!is.null(predictedData))
     data <- findModelOutlier(data = data,
                              predictedData = predictedData$observations,
                              yNames = ySelection$colNames,