+++ R/03-downUploads.R
@@ -8,7 +8,8 @@ downUploadsUI <- function(id, title) {
            value = id,
            fluidRow(
              sidebarPanel(
-               style = "position:fixed; width:20%; max-width:350px; overflow-y:auto; height:88%",
+               style = "position:fixed; width:23%; max-width:500px; overflow-y:auto; height:88%",
+               width = 3,
                selectInput(
                  ns("activePlot"),
                  label = "Select a saved plot",
@@ -85,8 +86,14 @@ downUploads <-
     # render plot ####
     values <- reactiveValues(plot = NULL)
     output$styledPlot <- renderPlot({
+      validate(
+        need(input$activePlot, "Select a plot ...")
+      )
       req(input$activePlot)
-      req(activePlotValues$predictedData$evenlyOnX)
+      validate(
+        need(!is.null(activePlotValues$defaultXRange), "Data not valid ...")
+      )
+
       makeSinglePlot(
         reactiveValuesToList(activePlotValues),
         reactiveValuesToList(activePlotStyle)
@@ -110,21 +117,21 @@ downUploads <-
       uploadedNotes = uploadedNotes
     )
 
-    dataFun <- reactive({
-      req(input$activePlot)
-      req(activePlotValues$modelData)
-      function(xVar, quantile) {
-        savedData <-
-          predictPipe(
-            plotRModel = activePlotValues$modelData$modelOutput,
-            xCol = activePlotValues$prepData$X,
-            xVar = xVar,
-            yName = activePlotValues$ySelection$colNames$colName1,
-            quantile = quantile
-          )
-        return(savedData)
-      }
-    })
+    # dataFun <- reactive({
+    #   req(input$activePlot)
+    #   req(activePlotValues$modelData)
+    #   function(xVar, quantile) {
+    #     savedData <-
+    #       predictPipe(
+    #         plotRModel = activePlotValues$modelData$modelOutput,
+    #         xCol = activePlotValues$prepData$X,
+    #         xVar = xVar,
+    #         yName = activePlotValues$ySelection$colNames$colName1,
+    #         quantile = quantile
+    #       )
+    #     return(savedData)
+    #   }
+    # })
 
     #callModule(dataExport, "exportData", dat = dataFun, filename = "modelData")
     callModule(plotExport, "export", reactive(values$plot))